Fix Accordion in playlists by moving content in Playlist.js back into PlaylistNavigation.js
