--- src/components/CoursePage/Playlist.scss
@@ -1,3 +1,0 @@
-.link {
-  cursor: pointer;
-}
