+++ src/components/CoursePage/PlaylistNavigation.js
@@ -2,21 +2,44 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import Accordion from 'react-bootstrap/lib/Accordion';
-import Playlist from './Playlist';
+import Panel from 'react-bootstrap/lib/Panel';
+import Badge from 'react-bootstrap/lib/Badge';
+import ListGroup from 'react-bootstrap/lib/ListGroup';
+import LessonItem from './LessonItem';
 import {getTranslator} from '../../selectors/translate';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import styles from './PlaylistNavigation.scss';
-import {getPlaylistsForCourse} from '../../resources/playlists';
+import {getPlaylistsForCourse, getPlaylistLessons, getPlaylistTitle} from '../../resources/playlists';
+import {areAllLessonsInPlaylistTranslated} from '../../resources/utils/playlistLessons';
 
-const PlaylistNavigation = ({t, course}) => {
+const PlaylistNavigation = ({course, language, t}) => {
   const playlists = getPlaylistsForCourse(course);
-  // TODO: Fix Accordion, seems like the current implementation got broken when upgrading to react-bootstrap 0.31.1
+
+  const playlistListItems = playlists.map(playlist => {
+    const lessons = getPlaylistLessons(course, playlist);
+    const title = getPlaylistTitle(course, playlist, language) || t('coursepage.missingtitle');
+    const header = <h4 role="presentation">
+      <Badge pullRight>{lessons.length}</Badge>
+      <span className={styles.link}>{title}</span>
+    </h4>;
+    return (
+      <Panel key={playlist} {...{header}} eventKey={playlist}>
+        <ListGroup fill>
+          {areAllLessonsInPlaylistTranslated(course, playlist, language) ?
+            lessons.map(lesson => <LessonItem key={lesson} {...{course, lesson, language}}/>) :
+            <span>{t('coursepage.lessonsnottranslated')}</span>
+          }
+        </ListGroup>
+      </Panel>
+    );
+  });
+  
   return (
     playlists.length > 0 ?
       <div className={styles.container}>
         <h3>{t('coursepage.lessoncollections')}</h3>
         <Accordion>
-          {playlists.map(playlist => <Playlist key={playlist} {...{course, playlist}} />)}
+          {playlistListItems}
         </Accordion>
       </div> :
       null
@@ -28,10 +51,12 @@ PlaylistNavigation.propTypes = {
   course: PropTypes.string.isRequired,
 
   // mapStateToProps
+  language: PropTypes.string.isRequired,
   t: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = (state) => ({
+  language: state.language,
   t: getTranslator(state),
 });
