+++ src/components/CoursePage/PlaylistNavigation.scss
@@ -1,3 +1,7 @@
 .container {
   overflow: hidden; // Avoid margin collapse
 }
+
+.link {
+  cursor: pointer;
+}
